fix: mock embeddings service via DI, fix message_store timestamp bug
- Create test/mock_embeddings_service.c as link-time replacement for
  the real embeddings service. No HTTP, no env vars — calls
  mock_embeddings_get_vector() directly.

- Fix embeddings abstraction leak: add embeddings_service_free_embedding()
  so callers never reach through to the low-level embeddings.h. Update
  document_store.c and pdf_tool.c to use it.

- Delete dead code: lib/util/pdf_processor.c/.h (nobody calls it).

- Fix message_store timestamp bug: structs stored seconds (time_t) but
  DB stores millis, causing channel_receive to miss messages published
  in the same millisecond. Changed all timestamp fields to int64_t,
  removed /1000 truncation in mappers, mark-read now uses last message
  timestamp instead of wall clock.

+++ lib/tools/messaging_tool.c
@@ -7,6 +7,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
+#include <inttypes.h>
 #include <pthread.h>
 
 static char* g_agent_id = NULL;
@@ -265,11 +267,11 @@ int execute_check_messages_tool_call(const ToolCall *tool_call, ToolResult *resu
         }
 
         p += snprintf(p, required_size - (p - response),
-                     "{\"id\": \"%s\", \"sender\": \"%s\", \"content\": \"%s\", \"created_at\": %ld}",
+                     "{\"id\": \"%s\", \"sender\": \"%s\", \"content\": \"%s\", \"created_at\": %" PRId64 "}",
                      msgs[i]->id,
                      msgs[i]->sender_id,
                      escaped_contents[i] ? escaped_contents[i] : "",
-                     (long)msgs[i]->created_at);
+                     msgs[i]->created_at);
         free(escaped_contents[i]);
     }
     free(escaped_contents);
@@ -566,12 +568,12 @@ int execute_check_channel_messages_tool_call(const ToolCall *tool_call, ToolResu
         }
 
         p += snprintf(p, required_size - (p - response),
-                     "{\"id\": \"%s\", \"channel\": \"%s\", \"sender\": \"%s\", \"content\": \"%s\", \"created_at\": %ld}",
+                     "{\"id\": \"%s\", \"channel\": \"%s\", \"sender\": \"%s\", \"content\": \"%s\", \"created_at\": %" PRId64 "}",
                      msgs[i]->id,
                      msgs[i]->channel_id,
                      msgs[i]->sender_id,
                      escaped_contents[i] ? escaped_contents[i] : "",
-                     (long)msgs[i]->created_at);
+                     msgs[i]->created_at);
         free(escaped_contents[i]);
     }
     free(escaped_contents);
